Did changes to the models to allow to connection to my Azure Database

+++ SuperStore P3/SuperStore P3/Models/Customer.cs
@@ -12,7 +12,7 @@ namespace Models
         }
 
         [DisplayName("Customer ID")]
-        public int CustomerId { get; set; }
+        public short CustomerId { get; set; }
         [DisplayName("Customer Title")]
         public string? CustomerTitle { get; set; }
         [DisplayName("Customer Name")]
