+++ SuperStore P3/SuperStore P3/Models/Order.cs
@@ -12,11 +12,11 @@ namespace Models
         }
 
         [DisplayName("Order ID")]
-        public int OrderId { get; set; }
+        public short OrderId { get; set; }
         [DisplayName("Order Date")]
         public DateTime OrderDate { get; set; }
         [DisplayName("Customer ID")]
-        public int CustomerId { get; set; }
+        public short CustomerId { get; set; }
         [DisplayName("Delivery Address")]
         public string? DeliveryAddress { get; set; }
