+++ SuperStore P3/SuperStore P3/Models/OrderDetail.cs
@@ -7,15 +7,15 @@ namespace Models
     public partial class OrderDetail
     {
         [DisplayName("Order Details ID")]
-        public int OrderDetailsId { get; set; }
+        public short OrderDetailsId { get; set; }
         [DisplayName("Order ID")]
-        public int OrderId { get; set; }
+        public short OrderId { get; set; }
         [DisplayName("Product ID")]
-        public int ProductId { get; set; }
+        public short ProductId { get; set; }
         [DisplayName("Quantity")]
         public int Quantity { get; set; }
         [DisplayName("Discount")]
-        public double? Discount { get; set; }
+        public int? Discount { get; set; }
 
         public virtual Order Order { get; set; } = null!;
         public virtual Product Product { get; set; } = null!;
